requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai>=1.3.5
python-dotenv==1.0.1
pydantic>=1.10.0,<3.0.0
pytest==8.2.0
pytest-asyncio==0.23.6
